backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pypdf
fpdf2
python-dotenv
